+++ src/core/plugin/index.ts
@@ -1,4 +1,4 @@
-import { Analytics } from '../../analytics'
+import { Analytics } from '../analytics'
 import { Context } from '../context'
 
 interface PluginConfig {
