+++ src/node/__tests__/node-integration.test.ts
@@ -1,4 +1,4 @@
-import { AnalyticsNode } from '../node'
+import { AnalyticsNode } from '../..'
 
 const writeKey = 'foo'
