+++ src/plugins/ajs-destination/__tests__/index.test.ts
@@ -3,12 +3,12 @@ import jsdom from 'jsdom'
 import { mocked } from 'ts-jest/utils'
 import unfetch from 'unfetch'
 import { ajsDestinations, LegacyDestination } from '..'
-import { Analytics } from '../../../analytics'
+import { Analytics } from '../../../core/analytics'
 import { LegacySettings } from '../../../browser'
 import { Context } from '../../../core/context'
 import { Plan } from '../../../core/events'
 import { tsubMiddleware } from '../../routing-middleware'
-import { AMPLITUDE_WRITEKEY } from '../../../__tests__/test-writekeys'
+import { AMPLITUDE_WRITEKEY } from '../../../test-helpers/test-writekeys'
 import { PersistedPriorityQueue } from '../../../lib/priority-queue/persisted'
 
 const cdnResponse: LegacySettings = {
