+++ src/plugins/ajs-destination/index.ts
@@ -5,7 +5,7 @@ import {
   SegmentEvent,
 } from '@/core/events'
 import { Alias, Facade, Group, Identify, Page, Track } from '@segment/facade'
-import { Analytics, InitOptions } from '../../analytics'
+import { Analytics, InitOptions } from '../../core/analytics'
 import { LegacySettings } from '../../browser'
 import { isOffline, isOnline } from '../../core/connection'
 import { Context, ContextCancelation } from '../../core/context'
