+++ src/plugins/analytics-node/__tests__/index.test.ts
@@ -1,8 +1,8 @@
 const fetcher = jest.fn()
 jest.mock('node-fetch', () => fetcher)
 
-import { Analytics } from '../../../analytics'
-import { AnalyticsNode } from '../../../node'
+import { Analytics } from '../../../core/analytics'
+import { AnalyticsNode } from '../../..'
 
 const myDate = new Date('2016')
 const _Date = Date
