+++ src/plugins/legacy-video-plugins/index.ts
@@ -1,4 +1,4 @@
-import { Analytics } from '../../analytics'
+import { Analytics } from '../../core/analytics'
 
 export async function loadLegacyVideoPlugins(
   analytics: Analytics
