+++ src/plugins/page-enrichment/__tests__/index.test.ts
@@ -1,4 +1,4 @@
-import { Analytics } from '@/analytics'
+import { Analytics } from '@/core/analytics'
 import { pageEnrichment, pageDefaults } from '..'
 
 let ajs: Analytics
