+++ src/plugins/segmentio/__tests__/index.test.ts
@@ -2,7 +2,7 @@ import assert from 'assert'
 import { mocked } from 'ts-jest/utils'
 import unfetch from 'unfetch'
 import { segmentio, SegmentioSettings } from '..'
-import { Analytics } from '../../../analytics'
+import { Analytics } from '../../../core/analytics'
 import { Plugin } from '../../../core/plugin'
 import { pageEnrichment } from '../../page-enrichment'
 import cookie from 'js-cookie'
