+++ src/plugins/segmentio/__tests__/normalize.test.ts
@@ -2,7 +2,7 @@ import assert from 'assert'
 import cookie from 'js-cookie'
 import { SegmentioSettings } from '..'
 import { normalize } from '../normalize'
-import { Analytics } from '../../../analytics'
+import { Analytics } from '../../../core/analytics'
 import { SegmentEvent } from '../../../core/events'
 import { JSDOM } from 'jsdom'
 import { version } from '../../../generated/version'
